use tokenizer in place of TextVectorization layer
Keras has a pre-built textvectorization layer. this seems nice when you have data which is already formatted in the TFDS standard, but it becomes difficult to work with if your data and labels are in another format. It seems one would have to write code to re-format the data into TFDS to do what is essentially just sentence tokenization.